Remove left-overs of an old implementation for symbolic values
This was based on a library named symja/sympy
and was removed in r15917 (b9816264).

In particular, it is nice that we can remove meconsole010.jar,
for which we do not know the license
and which bundles a lot of classes (it is 4 MB large!)
that we do not want to use,
but where accidental usages happen easily
(cf. the previous few commits that cleaned up such references).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@33668 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c