Convert a few txt files to Markdown
There is no disadvantage in doing so,
and it can be nicer to read.
Furthermore, the style of our txt files
(for example, how commands were indicated) was inconsistent.

Besides reformatting, small updates were done
like fixing references to renamed files,
but no major changes.

While we are at it, standardize names of readmes to README.md.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@33674 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c