Add machine-readable license declaration for ltl3ba
This program is under GPL, so we add special notes to our readme
to clarify the situation.

Furthermore, to fulfill the GPL requirements we either need to provide
sources together with the binary, or provide a download possibility.
It is easier to just commit the sources, so we do this.
They are added to this directory (and not ../sources/)
such that they are also added to the CPAchecker release archives.
The sources of BuDDy are needed because our ltl3ba binary
contains BuDDy.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@33683 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c