Add licenses and copyrights of all Ivy-managed runtime dependencies
We do not actually need to check this in,
but we need to ship this in the CPAchecker release archives,
and the easiest way to do this is to declare the licenses
in a way that we can use the reuse tool to manage them
and check everything in.

For non-runtime dependencies this is not necessary
because we do not bundle them with CPAchecker.

Unfortunately, most Java packages do not correctly declare their
licenses and copyrights in a machine-readable way,
so we cannot automate this for now.
Whenever we change the list of runtime dependencies,
we need to update the list in lib/java-runtime-licenses.txt manually.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@33685 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c