Set immediate dominator of start node to UNDEFINED
idom := immediate dominator
The iterative dominance computation algorithm requires setting the idom of the start node to
the start node itself. The start node cannot have an idom, so the public methods in DomTree
hide the fact that the start node is its own idom: every node that is its own idom (which is
impossible) is handled as if its idom is UNDEFINED.
It's easier to just change the internal representation, which makes handling this special case
unnecessary. With this commit, the start node's idom is set to UNDEFINED after the iterative
dominance computation has finished. This change shouldn't cause any externally observable
differences.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/program-dependence-graph@33687 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c