ASTConverterTest: Fix tests crashing when same JClass was instantiated twice
As Junit creates a new instance of the test class each time it executes a test of said class, we ran into an assertion error for JClassType. The assertion makes sure that the parent of JClassType does not have already a child which is equal to the JClassType to be added.
By initiating our JClassType once in the init method with @BeforeClass, we create an instance of our object once, which is then shared across all instances of our test class.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/java-development@33707 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c