ASTConverter can now handle arrays in instance of expressions
JRunTimeTypeEqualsType: Now accepts JReferenceType as typeDef instead of JClassOrInterfaceType. This is added because we need to handle ArrayType, which is part of ReferenceType.
ASTConverter: instanceof expressions can now handle arrays
Implementation before only handle JClassOrInterfaceType. But according to java language spec:
"The type of the RelationalExpression operand of the instanceof operator must be a reference type or the null type, or a compile-time error occurs."

We replace reference to JClassOrInterfaceType with its superclass JReferenceType and handle special cases for arrays.
RTTTransferRelation has references to code we changed. As it is not in scope to handle this test, the test is only fixed so it compiles and a TODO comment is added.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/java-development@33708 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c