Declare copyright for CongruenceCPA, FormulaSlicingCPA, and PolicyCPA
The same for related files.
These files were contributed by a developer working at the
University of Grenoble, so the copyright belongs to the latter
as confirmed by David Monniaux on 2017-02-01 via mail.

We do not need to declare every copyright holder
(just keep copyright notices that have been added)
but in easy cases like this we can add the mention to make it clear.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@33722 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c