Remove a few external test suites from the repo
These files had no license declaration
and were never used since they were added years ago.
For most of them, a license could not even be easily researched and
added because the original test suite already violates the license
of the projects from which code was copied into the test suite.

Furthermore, many of these programs were preprocessed with CIL
before they were added to the repo, which we would want to avoid
if we started using them in the future.
And of course, today we would add such files to the SV-Benchmarks repo
instead of the CPAchecker repo.

Removed test suites:
- VeriSec: http://se.cs.toronto.edu/index.php/Verisec_Suite
  (we only had a subset of actual suite;
   contains redistributed code without copyright notice;
   preprocessed with CIL instead of original source)

- CPBM: http://www.cprover.org/software/benchmarks/
  (we only had a subset of actual suite;
   preprocessed with CIL instead of original source)

- Olden benchmark suite: https://www.martincarlisle.com/olden.html
  (we only had a subset of actual suite;
   preprocessed with CIL instead of original source)

- ECA: http://www.rers-challenge.org/2012/index.php?page=problems
  (was moved to SV-Benchmarks in the past)

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@33738 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c