Add copyright header for a few files copied from SV-Benchmarks
minepump has copyright by University of Passau
(not yet documented in SV-Benchmarks)
because it was written by an employee of the university.
rule60_list2.i is from LDV, so we can use it under the usual license
and copyright as all their other contributions to CPAchecker.
The remaining files have proper copyright documentation in
SV-Benchmarks.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@33741 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c