Declare license of test programs derived from the Linux kernel.
Most of these were modified by others, but the license of course is
still the GPL-2.0.
Specifically for programs for LDV it is enough to declare
the copyright of the kernel authors.

Because of the number of files and because we do not redistribute them
in release archives, we can use .reuse/dep5 as a shortcut for the
declaration. But we do so only for files that are in clearly separate
directories with no risk of confusion due to files that are added later.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@33742 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c