Declare copyright for files from CBMC regression-test suite
Source: https://github.com/diffblue/cbmc/tree/develop/regression/

These are many files copied from another source,
we do not redistribute them in our release archives,
and they are in a clearly separate directory,
so it is enough to specify the license in .reuse/dep5.

A better long-term solution might even be to use
(a fork of) the CBMC repository for storing these files.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@33746 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c