Remove not working test for fullscreen requests
The removed test only checks the return value of the functions that request a change to fullscreen and not whether the browser actually is in the fullscreen afterwards. Due to security reasons it is difficult to reproduce a fullscreen request with automated tests and therefore the whole test was removed.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/report-js-test-improvements@33823 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c