MPI Portfolio: Overhaul complete class
I have made most of the changes several weeks ago, and as there are
so many, it's not really viable at this point to split them into
partial commits anymore (hence this large commit).

The class is now fully functional, but my goal is to refactor it in
a more OO-way as soon as I have the time for it.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@33958 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c