Convert PolicyIterationTest from JUnit to BenchExec
It is an integration test, so let's use our existing setup for
integration tests.
Because of a solver regression, one test currently times out,
and this is better handled in our integration-test setup.

The config is used is now the main config, not specific test configs.
This is an advantage because it will also detect broken changes in the
main config files.
However, we specify the same individual config options for some tests
as it was done previously.

For some overflow tasks, the task-definition files contained the wrong
property, so this also needs to be fixed
(with the expected verdict taken from the JUnit test).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@33972 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c