Improve cases where a new specification file is added to a Specification
This happens for example in counterexample checks,
where we want to use the original specification plus an additional
automaton.

In most cases, a fresh Specification instance was created from scratch,
which meant that all automata that were already present in the original
specification instance had to be parsed again.
Now we parse only the new automata.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@33986 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c