Always store properties in Specification instance
This fixes a bug from r23727 (09849f3e),
which added properties to the Specification instance,
but only if there is at least one automaton
(i.e., if specFiles is not empty).

This problem was noticed before and fixed for coverage properties
in r29908 (9311b326). The current change fixes it for all properties,
so we can remove the code that was added for fixing it for coverage.

This bug was not noticed because it never had any effect for
verification properties:
For verification properties except termination the code in CPAMain
enforces that an automaton is added to specFiles,
so the condition triggered only for termination.
But for termination, code in CoreComponentsFactory would ensure
that a Specification instance with the termination property
is always created.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@33995 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c