Fix a bug in CPAchecker class when analyzing a result
The result of getViolatedProperties() on a target state was allowed to
be empty according to its documentation,
but the CPAchecker class checked only whether the reached set has at
least one target state with a non-empty set of violated properties,
instead of whether the reached set has a target state.
Now we do the latter.

Of course, precise information about the violated property for the user
is only available if the target states return violated properties,
but at least we do not return UNKNOWN anymore for a reached set with
target states.

This fixes this regression in PCC resulting from r33981 (4e551e2d):
https://buildbot.sosy-lab.org/cpachecker/results/nightly-pcc-result-check-partial-reached-predicate/00143.nightly-pcc-result-check-partial-reached-predicate.diff.html#/table

Also update some JavaDoc to make it more clear.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@34004 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c