Update SoSy-Lab Java Common
This fixes a bug in injecting configuration options
with relative paths if the option name was prefixed in the config file.
For example, when handling the option "TargetCheck.cpa.automaton.inputFile"
for a CPA instance with the name "TargetCheck",
a relative path was treated as relative to the current working directory
instead of relative to the directory of the config file.
Options without prefixes (e.g., "cpa.automaton.inputFile")
were handled correctly.

This occurs in some many config files that I wonder why nobody has
detected and complained about this before?

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@34005 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c