Change how termination_as_reach.spc is added in our termination analysis
Our TerminationAlgorithm needs a specific automaton
(which is written down in config/specification/termination_as_reach.spc)
to be present in the tree of CPAs during the analysis.
This automaton marks certain dummy CFA nodes of the termination
algorithm as targets, so the algorithm and the automaton depend on each
other.

So far, some code in CoreComponentsFactory ensured that whenever the
termination algorithm was used, this specific automaton was added to the
Specification instance, and thus afterwards instantiated as CPA and
added to the tree of CPAs (like any automaton given as specification).
Furthermore, this code also ensured that no other automata
(except for an optional witness) were contained in the specification
when termination analysis was enabled.
Both parts are removed now.

First the explanation of the former part of the change:

The special handling of adding termination_as_reach.spc is removed,
and instead this automaton needs to be added as a CPA in the config files
for termination (and so it is done in the provided config files).
This is of course much nicer from a coupling-related point of view.
Previously CoreComponentsFactory needed to know this specific
implementation detail of TerminationAlgorithm and about this specific
automaton file, because it needed to ensure that the Specification
instance was adjusted appropriately.
In those configuration files that explicitly enable the termination
algorithm it is however perfectly fine to need some additional
configuration options required by the termination algorithm
(this is just like the PredicateCPA being needed when BMC is used,
 and even the TerminationAlgorithm already requires explicitly adding
 the TerminationCPA in the config file, so we can add other CPAs as well).
It is also a little bit more flexible.
Developers who wish to experiment with a modified version of the
automaton can now simply configure this instead of needing to change
an unrelated class.

Furthermore, although this automaton is actually used to define which
abstract states are target states, it is not really a specification as
it is in the usual sense.
The specification is to check for termination,
and this automaton is just part of one technical solution implementing
this, it cannot be used to tell CPAchecker to check for termination.
So adding it to the Specification instance can be seen as questionable,
it feels more like an auxiliary CPA.
The new way of configuring it reflects that.
This might be an advantage if other code looks at the Specification
instance and does not expect to see a technical implementation detail in it.

This part of the change has two visible effects:
- Users/developers who write their own termination-related config file
  need to add this automaton. Documentation is added, and most people
  should copy or include our provided config file, so this should not be
  a problem.
- Because the Specification instance is not changed anymore,
  we also do not add the termination property anymore.
  This has only the effect that termination witnesses will now contain
  the termination property only if the user specifies that property
  explicitly on the command line with "-spec .../termination.prp".
  Previously a termination witness would always contain the termination
  property.
  The new behavior is the same as for all other properties that
  CPAchecker supports and thus should also not be a problem.
  (If it were, we should probably find a general solution for all
   properties.)

Second the explanation of the latter part of the change:

The check that forbids everything in the Specification instance
when using TerminationAlgorithm except for
termination_as_reach.spc and an optional witness
is deleted without replacement.

This check was too strict and causing problems:
Because for witnesses only the file extension ".graphml" was allowed,
compressed witnesses or renamed witnesses were incorrectly forbidden.
This even happens in our own integration tests:
https://buildbot.sosy-lab.org/cpachecker/results/nightly-termination-validation/00108.nightly-termination-validation.witnessValidation.results-validated.table.html#/table
(all rows with status "unknown" in the validation run set).

In principle we could fix the check and allow renamed witnesses
(and move the check into TerminationAlgorithm, where it belongs
 and where it could be now).
However, the code that actually parses automata and witnesses
does not care at all about file names, it just checks the content.
So even then it would be possible to trick the termination analysis
into allowing an arbitrary automaton by just naming it
"something.graphml".
And additionally it is possible to simply add instances of the
ControlAutomatonCPA to the list of CPAs and load arbitrary automata
(as we do now for loading termination_as_reach.spc!),
or to encode an arbitrary automaton in the witness format.

So the existing check does not really prevent anything,
it just catches the most obvious cases like running
"-terminationAnalysis -spec ErrorLabel".
However, these are certainly not common mistakes,
because for termination analysis one would either use
"-terminationAnalysis", which does not need a separate specification at all,
or a configuration like "-default" or "-svcomp20",
in which the termination analysis is triggered only
if "-spec .../termination.prp" is passed in addition anyway.

Additionally, because we see CPAchecker as a platform and framework,
we typically want to avoid hard-coding such relations
between components like a certain specification and a certain algorithm.
After all, we do not know if some use case for this combination might happen to exist.
For example, we do not have code that prevents running
"-predicateAnalysis -spec termination_as_reach",
so why do this differently in the termination analysis?

So let's just remove this broken check completely.

In summary, these changes remove tight coupling between unrelated
components and make the configuration of the termination analysis
be much more similar to how other analyses are configured,
with user-visible changes only affecting rare cases.

There is one additional detail necessary to make this work:
Instances of ControlAutomatonCPA support only a single automaton
(includes of another automaton are not possible),
so we need to remove the existing include in termination_as_reach.spc
and define two CPA instances.

And for those of you who have read this far
(yes, it is the longest commit message in this repo until now),
there is one more positive side effect:
Previously when validating termination witnesses,
the witness would have been parsed at least three times:
once regularly, and twice when creating the two adjusted Specification
instances in CoreComponentsFactory. This is now avoided.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@34011 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c