Add possibility to override unsupported functions of predicate analysis
In r32980 (80e35da7) we added a list of functions that let the
predicate analysis abort immediately because we know they are
unsupported and wrong results can be produced (example: memset).

However, there are users who would like to override this
because they know that in their case such calls are not a problem
(example: LDV).

So let's add an option that allows this.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@34133 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c