Developing a new style of merge for PredicateMergeOperator
The current merge operation does not merge abstraction states.
In fact, two abstraction states can be merged if their preceeding abstraction states are identical.
This commit adds an option to PredicateCPA to toggle such behavior of the merge operator.
This option also hides the unfinished code so that predicate analysis will not be affected.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@34140 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c