Modify the implementation of merge between two abstraction states
The original implementation was not even called because `AbstractionState` implemented `NonMergeableAbstractState`. This has been changed in this commit. Moreover, the original implementation used ARGCPA to retrieve the preceding abstraction states, which resulted in `ClassCastException` as we only have `PredicateAbstractState` in the merge operator. To solve this, a field to remember the previous abstraction state has been added.

The logic to record the previous abstraction state is summarized as follows:
1) The initial state has itself as previous abstraction state
2) A successor has its parent as previous abstraction state if its parent is an abstraction state; otherwise its previous abstraction state is the same as that of its parent
3) An abstraction state created by prec operator has the same previous abstraction state as that of the intermediate state from which it was created
4) States created by merge operator have the same previous abstraction states as that of the second element

For other code which also creates predicate abstraction states, they have not been modified yet. That is, they will create states whose fields to store previous abstraction states are null. This should be okay as long as this field is not accessed.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@34185 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c