Change the field storing previous AbstractionState to final
This change results in overloaded methods and additional constructors.
When callers do not provide a previous abstraction state,
the original methods/constructors are used and this field will be initialized to null.

The previous abstraction state of the initial state is changed to null,
as this is more consistent with the meaning of the field.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@34193 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c