Add option to check for missed witness invariants
This commit adresses issue #736. There are cases where we want to
detect whether there are invariants in the witness that are not
actually being checked by the witness validation. Most of the time
this will be cause by a bug, but in theory there are cases where
this is perfectly fine (though useless).

Currently the new option witness.debug.checkForMissedInvariants
will cause that a CPAException is thrown in such a case. After some
testing, we might default to always check this, but output a WARNING
instead of throwing an exception. We could still allow to throw the
exception based on the configuration option. Another way would be
to have a dedicated configuration/analysis that checks for this.
It should be quick and would enrich the integration tests to find
cases where we do not make use of our own invariants.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@34238 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c