Fix incomplete removal of SSA indices after addValueImportConstraints
Background: When a variable is first used, the predicate analysis
knows it is not aliased and uses an optimized formula encoding as a
simple variable. If later on the address of the variable is taken,
it is encoded differently in the formula and the previous SMT variable
must not be used anymore.
This conversion is handled among other code by
CToFormulaConverterWithPointerAliasing.addValueImportConstraints.
To signal that the variable should not be used anymore,
we remove it from the SSAMap
(in CExpressionVisitorWithPointerAliasing.visit(CUnaryExpression)),
but this did not handle structs on the stack.
For these only the variable for the whole struct was removed from the
SSAMap, not the variables that are used to encode individual fields.

This caused problems later on, e.g., in AssignmentToPathAllocator,
as described in #739.

This complicates handling of SSAMap indices during merging of
PointerTargetSets (which also needs to call addValueImportConstraints),
but it is also correct to do this there
(otherwise merging of PathFormula's would reintroduce SSA indices
for variables for which the index was removed in one of the branches).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@34292 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c