Make code related to CFAEdgeWithAssumptions more deterministic
This affects code that builds the list of assumptions that will be
stored in a CFAEdgeWithAssumptions, and the code that assigns
CFAEdgeWithAssumptions to ARGStates.

The order of assumptions influences the output of witnesses and
counterexamples, so it should be deterministic.
Furthermore, the easiest way to make it deterministic is to use
immutable collections and their builders consistently,
which also increases the overall code quality.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@34299 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c