Remove some useless defensive copying in AssignmentToPathAllocator
There is a loop, and some collections were copied in each iteration,
but this is useless, because no references to these collections are
stored anyway. Only immutable copies of these collections will end up
stored somewhere.

Furthermore, there are actually four collections that are used in the
same manner, and only for three of them the defensive copies were made
("variables" is the fourth collection).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@34302 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c