Refactor SLAB's EdgeSet a little bit
It had a constructor that took a passed set and directly assigned it to
a field, without making a copy.
This is typically error prone because callers do not expect it,
in particular not because the other constructor of EdgeSet creates a copy.
If no copy is intended, it is better to create a factory method
with a name that makes this clear
(e.g., "public static EdgeSet wrapping(Set<CFAEdge>)").
Furthermore, the one caller of this used it to create an EdgeSet
wrapping an immutable collection, although EdgeSet is mutable and
intended to be used with mutable collections,
so this could also be confusing.
And this caller did not even use any of the features of EdgeSet,
a regular ImmutableSet<CFAEdge> would be just as good.

So we can cleanup all of this by using ImmutableSet<CFAEdge> in the
caller and making the constructor properly copy the passed set.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@34309 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c