Introduce two checks to make the code more robust against wrong results
The first one checks if there are target states in ARG.
Theoretically, if there are no target states in ARG, the program should be safe.
However, when single-loop transformation is used,
there are cases where no target states exist, but the program is unsafe.
To avoid these wrong TRUEs, this check will throw exceptions if no target states exist in ARG.

The second one checks if the collected formulas are insane.
There are cases with single-loop transformation where the conjunction of prefix and loop formulas is UNSAT.
This UNSAT results in a false interpolant, and reaches a fixed point.
To avoid these wrong TRUEs, this check will throw exceptions if the conjunction is UNSAT.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/interpolation-model-checking@34339 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c