Do not throw exceptions when there are no target states
There are indeed cases where no target states exist in ARG.
For example, a loop-free program will not have target states for k>1,
if the target states at k=1 are unreachable.
These programs should be correctly proved instead of being treated as exceptions.

If no target states are found when k=1, we keep unrolling.
If no target states are found when k>1, we conclude the program to be safe.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/interpolation-model-checking@34343 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c