Refactor reuse of interpolation environment
When reusing a single interpolation environment for all interpolation
queries, we need to store the formulas from the last query such that we
can determine which parts of the solver stack we need to clean up for
the next query.
So far we also stored the AbstractState that belonged to each part,
now we avoid this.
This is of course more efficient (because some these states will be
unused after the refinement and can be garbage collected),
but there is one more effect.
When creating the new list of (formula, state, formula id)
for the next query, we would so far partially store old states in it.
They should be the same as the new states,
but it still seems better to use the proper new states
even for those parts of the interpolation query where the formulas are
still the same and the solver stack is reused.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@34443 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c