Improve defaults for a few interpolation-related config options
This is based on the Master's thesis
"Heuristics for Effective Predicate Refinement in CPAchecker" from 2019
and I confirmed the most important findings.

Setting cpa.predicate.refinement.cexTraceCheckDirection=ZIGZAG
together with cpa.predicate.refinement.incrementalCexTraceCheck=true
gives a nice increase in number of solved tasks.
This is understandable because it forces the solver to pick interpolants
for the start or end of the counterexample
(which are typically more relevant) with less chance of getting
interpolants for irrelevant loops somewhere in the middle.

Setting cpa.predicate.refinement.predicateBasisStrategy=SUBGRAPH
sometimes avoids throwing away already discovered predicates
if a later refinement has a cut point (pivot state) that is higher up in
the ARG than a previous refinement.
This can reduce the number of refinement.

Also update the description of these options.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@34450 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c