Add yet another fallback mechanism for interpolation queries
MathSAT interpolation queries sometimes fail
(e.g., with "can't build ie-local interpolant") but we have noticed that
MathSAT could solve the same query in different circumstances
(for example, without incremental SAT checks and if the formulas were
pushed in a different order).
This does not help always, but it does not cost us much in trying,
because in most configurations CPAchecker would just terminate with
UNKNOWN if we give up.
For configurations (i.e., in a sequential combination) where this is not
desired, we add an option.

So far the fallback changes two things:
- reverse order of pushed formulas
- inverted flag for incremental checks
  (i.e., non-incremental if first check was incremental and vice-versa)

These seem to be all options of InterpolationManager where changing them
makes sense. If we find more options that could have an influence on
this, we should add them.

Note that there is already a fallback mechanism available via
cpa.predicate.refinement.sequentialStrategy=FWD_FALLBACK.
However, this fallback only retries the actual getInterpolant calls
to the solver, and in my experiments did not provide any benefit.
In contrast, this new code uses a new prover environment and lets the
solver solve the formulas from scratch again before asking for
interpolants.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@34452 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c