Add yet another tool for checking our style guide
This is called Policeman's Forbidden API checker
and focuses on calls to unwanted APIs:
https://github.com/policeman-tools/forbidden-apis

Our existing tools do not cover this fully:
- Checkstyle cannot reliably detect calls to unwanted methods
  (cf.  https://github.com/sevntu-checkstyle/sevntu.checkstyle/issues/773).
- Refaster is great if there is an automatic replacement,
  but this is not always the case.
- Customizing SpotBugs would require writing Java code.

For example, Collections.shuffle() should only be called by using an
instance of Random with a fixed seed and passing it as second parameter
in order to avoid nondeterministic behavior of CPAchecker,
but this cannot be enforced well with our existing tools.

Another use case is lots of overloads (e.g., Lists.of()),
where writing a Refaster rule for each of them is not worth the effort,
but with this tool it is easy enough.

The new tool covers this case nicely and is easy to use and fast enough.
Forbidden classes and methods can be written down in a simple text file
(build/forbiddenapis.txt).
The error message is customizable for each case and will be shown as
text output, just like a compiler error.

The new check is also included in "ant standard-checks",
so developers who use this for checking their code before pushing
won't need to do anything in addition.
Of course, it is also added to the CI.

Suggestions for additions to the list of forbidden APIs are welcome!

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@34521 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c