Do not rely on the collected formulas for BMC and forward conditions
It brings no benefit to use the collected formulas for BMC and forward conditions.
Instead, we just trace back from the target or stop states, and build the formulas.

On the other hand, the collected formulas should only be used if the ARG satisfies:
1) No covered states exist and 2) Exactly one loop head ABS for every k.
We check these two requirements after each unrolling, when interpolation is enabled.
If the ARG does not meet the above requirements,
we disable interpolation, and roll the analysis back to plain BMC with LBE.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/interpolation-model-checking@34526 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c