Read property type from witness instead of from current analysis config
This reverts r26425 (febf5e09) as discussed in
https://gitlab.com/sosy-lab/software/cpachecker/-/commit/febf5e09fa82443feb8d93b5b42aab1a11436769#note_384250940

This information is used only for deciding whether invariants are
allowed in the given witness or not.
It makes more sense to base this decision on the specification that is
contained in the witness instead of what kind of specification is
configured for the current analysis.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@34531 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c