Add a never-stop operator for predicate CPA to avoid covered states
An abstract state is covered by another abstract state if
the former's abstraction formula is implied by the latter's.

In BMC with LBE, such covered states make the collected path formulas incomplete.
The ABS merge operator partially solved this problem for single-loop programs,
but for multi-loop programs, there are ABS states with different predecessors.
Therefore, they cannot be merge by the ABS merge operator.

On the other hand, in BMC the abstraction formulas are always true.
Checking that true implies true and therefore not adding a state into waitlist is not beneficial.

In BMC with whole program encoding, i.e., using blk-never,
there are no abstract states, so this was not a problem.

This never-stop operator will always add new states into waitlist.
It should not be used when there are no other termination mechanisms.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/interpolation-model-checking@34579 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c