Refine never-stop operator to respect ABS merge operator
Stopping only because both states are abstract will lose the chance to ABS merge them,
if they have the same preceding abstract state.

The refined condition for a non-stop is
1) Both e1 and e2 are abstract states, and
2) e1 and e2 have different preceding abstract state.

If the condition is true, the check of pair (e1, e2) will be skipped.
Otherwise, e1 will be handled in the same way as PredicateStopOperator.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/interpolation-model-checking@34583 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c