Perform interpolation only if there exists a target state in ARG
There are test cases with SLT where no target state exists when maxLoopIterations=2.
In such cases, interpolation will conclude the program to be safe.
However, reachable target states will appear if we keep unrolling.

This commit strengthens the precondition for interpolation such that
it is performed only if there exists a target state to avoid wrong TRUEs.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/interpolation-model-checking@34600 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c