Fix class class cast exception in AbstractExpressionValueVisitor#visit(JArraySubscriptExpression) in case a variable was given as array index
In case a variable was passed as JArraySubscriptExpression, inside the method, a Value.Unknownvalue was returned. This was tried to be cast to NumericValue, which resulted in a class cast exception.
Will now first check before casting to NumericValue. In case it is not a NumericValue, Value.Unknown will be returned.

This behavior can be checked via java/jbmc-regression/ArrayIndexOutOfBoundsException2 of the sv-benchmark suite

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/java-development@34706 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c