Fix very rare bug when strengthening ValueAnalysisTransferRelation
For program java/java-ranger-regression/apachecli_eqchk of the sv-benchmarks suite, ValueAnalysisTransferRelation had a missingInformationLeftJVariable. Current solution is rather a workaround because after long search, reason could not be found:
In case missingInformationLeftJVariable is null, we look up the name of the declaration in the previousEdge, reconstruct its name, look whether it is found in the constants and if it is found, set missingInformationLeftJVariable to this name.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/java-development@34736 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c