Add todo with reference to this commit message for debug information
The problem occurs with /sv-benchmarks/java/java-ranger-regression/apachecli_eqchk

The method ValueAnalysisTransferRelation#strengthen(RTTState rttState, CFAEdge edge) is called multiple times for variable longOptsLen in line 43. At the first call of the method, missingFieldVariableObject gets changed from true to false. Then, it jumps into this workaround.

It needs to be looked at whether a second call of this method is intended or if missingFieldVariableObject should not be set to false in some cases.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/java-development@34737 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c